requirements.txt
flask
boto3
pandas
requests
fpdf
python-dotenv